tests: Align integration tests
